refactor(layout): move snapshot to vmux_layout, classify tabs by URL
