+++ crates/vmux_layout/src/lib.rs
@@ -7,6 +7,8 @@
 pub mod event;
 pub mod protocol;
 pub mod reconcile;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod snapshot;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub mod chrome;
